refactor(types): split monolithic file into modular structure
Split the 1305-line types.rs into a well-organized directory:
- mod.rs (61 lines) - Module orchestration and re-exports
- person.rs (62 lines) - Person and PhoneNumber types
- room.rs (82 lines) - Room, RoomListParams, RoomType, SortRoomsBy
- organization.rs (98 lines) - Organization, Team, Catalog, Destination
- membership.rs (49 lines) - Membership and MembershipListParams
- message.rs (136 lines) - Message types and parameters
- device.rs (86 lines) - Device and Authorization types
- attachment.rs (44 lines) - Attachment and AttachmentAction
- event.rs (750 lines) - Event, Activity, GlobalId types with tests
- api.rs (76 lines) - Internal Gettable trait and ListResult

Benefits:
- Better code organization and maintainability
- Logical grouping of related types
- Easier to navigate and find specific types
- All tests passing (37/37)
- Preserved all functionality including tests
